+++ backend/notes/models.py
@@ -1,3 +1,12 @@
 from django.db import models
 
-# Create your models here.
+
+class Note(models.Model):
+    note = models.CharField(max_length=255)
+    description = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    created_by = models.ForeignKey('auth.User', on_delete=models.CASCADE,)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.title